fix socialIcons component
