--- components/layout/footer/children/types.ts
@@ -1,8 +1,0 @@
-export interface FooterProps {
-  socialIcons: SocialIcon[];
-}
-
-export interface SocialIcon {
-  iconSrc: string;
-  socialLink: string;
-}
